gql-weather/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Paul Bailey", email = "paul@neutron.studio"},
]
dependencies = [
    "django>=4.2.7",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "dj-database-url>=2.1.0",
    "httpx>=0.25.1",
    "honcho>=1.1.0",
    "setuptools>=68.2.2",
    "uvicorn[standard]>=0.24.0.post1",
    "whitenoise>=6.6.0",
    "psycopg[binary,pool]>=3.1.13",
    "graphene-django>=3.1.5",
    "django-filter>=23.4",
    "django-ws>=1.0.0",
    "graphene-luna>=1.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "BSDv3"}

[tool.pdm.scripts]
_.env_file = ".env"
